Handle cyclic routes when loading a bad catalogue
Try and generate the layouts when a catalogue is loaded.
Store layouts in the app instead of the frame.
Specify module separation distances in ModuleLayoutUtil
instead of having to pass them in.
Catch the exception when generating the layout, show a
warning, and delete the routes from the process.
Try and html-escape all names that get shown
in <html> blocks.